requirements.txt
awswrangler
pandas
boto3
mysql-connector-python
requests